+++ src/isisdl/__main__.py
@@ -5,7 +5,7 @@ import isisdl.compress as compress
 from isisdl.backend.request_helper import CourseDownloader
 from isisdl.utils import args, acquire_file_lock_or_exit, generate_error_message, subscribe_to_all_courses, unsubscribe_from_courses, install_latest_version, export_config, database_helper, \
     config, migrate_database
-from isisdl.config import init_wizard, config_wizard
+from isisdl.backend.config import init_wizard, config_wizard
 from isisdl.settings import is_first_time
 from isisdl.settings import is_online
 from isisdl.version import __version__
